[Game] The event window became active;
[Packets] Fixed Packets:
- SCDailyCountPacket;
- SCUnitStatePacket: added for Npc the output of the spawnerId field;

[Packets] Added new packets:
- CSGameEventPacket;
- SCAbilitySetAllInfoPacket;
- SCGameEventPacket;
- SCGameEvent2Packet;
- SCRestrictPacket;
- SCUiTextsKeyPacket;

+++ AAEmu.Game/Core/Packets/G2C/SCDailyCountPacket.cs
@@ -5,11 +5,11 @@ namespace AAEmu.Game.Core.Packets.G2C;
 
 public class SCDailyCountPacket : GamePacket
 {
-    private readonly int _totalCount;
-    private readonly int _dailyCount;
-    private readonly int _dailyMaxCount;
+    private readonly byte _totalCount;
+    private readonly byte _dailyCount;
+    private readonly byte _dailyMaxCount;
 
-    public SCDailyCountPacket(int totalCount, int dailyCount, int dailyMaxCount) : base(SCOffsets.SCDailyCountPacket, 5)
+    public SCDailyCountPacket(byte totalCount, byte dailyCount, byte dailyMaxCount) : base(SCOffsets.SCDailyCountPacket, 5)
     {
         _totalCount = totalCount;
         _dailyCount = dailyCount;
